feat: use chat box + remove ability to edit initial question
Previously, the ability to edit the initial question was favoured due to 2 reasons
1. This seemed to be more intuitive in informing the user that there is no chat history
2. Lower development effort in ensuring that the chat box component is always fixed at the bottom

However, due to feedback that it is not intuitive that the initial question can be edited,
this commit reverts back to using a chat box component. At the same time, this commit ensures
that the chat box component is always fixed at the bottom in the chat interface screen.