fix: sidebar component erroneously renders twice when switching between tabs
Previously, as the sidebar component renders twice, the loading indicator will appear
momentarily when switching between tabs.

This commit fixes the issue by making the queryKey a simple constant value. In doing so,
TanStack query is better able to manage query caching and will not perform a fetch whenever
the session is updated.

See here for more information on how the queryKey works:
https://tanstack.com/query/v5/docs/framework/react/guides/query-keys