fix: unable to click back when an existing stream is still ongoing
Before this commit, when a user submits a request and before the server returns the full response, the user is unable to click back to the home page to send a new query. The user can only do this when the server finishes streaming its response.

To fix this, we expose the cancelStream function in useStreamingQuery to the ChatContext so that whenever the handleBackClick function is invoked, any existing stream will be aborted.